Add ability to render md files and move posts over
This adds the ability to render markdown files for the site. It also
ports over all the old blog posts to this directory. The .gitignore file
makes it so that any files there aren't commited to the repo as they are
all tracked in _posts. By having the .gitignore file there though it
allows us to keep the posts directory tracked.